validator_experimentos_automaticos/validator_python/requirements.txt
requests==2.31.0
pytest==7.4.3
python-dotenv==1.0.0
pyyaml>=5.4.1
matplotlib>=3.4.3
numpy>=1.21.0
fastapi==0.109.0
uvicorn==0.27.0
python-socketio>=5.4.0
websockets>=9.1
scikit-learn>=0.24.2
opencv-python>=4.5.3 